Decouple Messages module from announcement and alert settings.
The Messages module defines methods for sending messages to various
locations and for creating the messages to send.  Move the logic to
determine whether to send the messages at all to the locations that
call these methods.

Also use the AceConsole-3.0 "Print" method instead of rolling our
own print function to prefix output with the name of the addon in
a pretty color. This makes the output look more consistent.