Allow user to filter debugging messages by level of detail.
Add a new debugging option "Debug level" (valid values are 1 to 3)
that specify the level of debugging detail that should be output.
The default level is 1 for just outputting any unknown loss of
control events. Higher levels show more and more details of
function and method calls invoked by the running addon.